requirements.txt
fastapi[standard]
psycopg2
Jinja2
python-dotenv
